feat: allow invoking cmake-ts as a library
