+++ src/lib.ts
@@ -1,1 +1,2 @@
-export * from "./config.js"
+export * from "./build.js"
+export type * from "./config.js"
